Moving the cross_platform.hpp header under include/
