+++ inavjaga.cpp
@@ -1,4 +1,4 @@
-#include "cross_platform.hpp"
+#include "include/cross_platform.hpp"
 #include "inavjaga.hpp"
 #include <algorithm>
 #include <future>
