+++ include/cross_platform.hpp
